Several feautures (#27)
* feat: gravity

* feat: speed

* fix: speed

* docs: add speed and gravity


* feat: hp regen

* feat: armor regen

* docs: remove extra semicolon

* fix: rename hp to Health

* docs: missing comma in config


* feat: fast defuse

* feat: fast plant

* comment debug logs

* introduce .editorconfig

* add TimeAfterRoundStart for fast plant and defuse

* docs: TimeAfterRoundStart

* prepare bhop config

* feat: bhop

* docs: bhop

* unhook CBaseEntity_TakeDamageOldFunc

* move times into Timers.cs

---------

+++ src/Config/VipGroupConfig/MiscConfig/ExtraJumpsConfig.cs
@@ -8,7 +8,7 @@ namespace Core.Config
         public int Amount { get; set; } = 0;
 
         [JsonPropertyName("VelocityZ")]
-        public double VelocityZ { get; set; } = 260.0;
+        public float VelocityZ { get; set; } = 260.0f;
 
         [JsonPropertyName("NoFallDamage")]
         public bool NoFallDamage { get; set; } = true;
